+++ acceptance-tests/page-objects/dashboard/components/dashboard-variable/dashboard-notebook-variable.component.ts
@@ -1,0 +1,18 @@
+import { Locator, Page } from "playwright/test";
+import { DashboardVariableBaseComponent } from "./dashboard-variable-base.component";
+
+export class DashboardNotebookVariableComponent extends DashboardVariableBaseComponent {
+    constructor(page: Page) {
+        super(page);
+    }
+
+    public get notebookVariableDropdown() {
+        // return this.page.locator('#react-select-6-input');
+        return this.page.locator('text=Select notebook').first();
+    }
+
+    async selectNotebookVariableDropdownOption(option: string): Promise<void> {
+        await this.notebookVariableDropdown.click({ force: true });
+        await this.page.getByText(option).click();
+    }
+}
